refactor(node): Move `EventEmitter` definition to `events.d.ts`
